CNN_MODEL
Poultry diseases including Coccidiosis, Salmonella, and Newcastle can lower chicken productivity if they are 
not detected early on. In this study, a 
Convolutional Neural Network based framework has been proposed to classify poultry diseases by distinguishing 
healthy and unhealthy fecal images. Unhealthy images can be a sign of the poultry diseases. The Image Classification 
dataset was used to train the framework, and it was discovered that it performed with an accuracy of 99.99%, 96.05%, 
93.23% on the training set, validation set, testing set respectively. When the proposed network's performance was 
evaluated against pre-trained models, it was discovered that the proposed model was unquestionably the best one for 
classifying chicken disease. This framework can beat resource-intensive machine learning methods due to the trained 
model's reduced weight and can be implemented with a small amount of memory and computational power.